refactor: extract image validation logic to shared utility
- Create imageUtils.ts with common image validation functions
  - ALLOWED_IMAGE_EXTENSIONS constant
  - ALLOWED_IMAGE_MIME_TYPES constant
  - getMimeTypeFromFilename() function
  - isAllowedExtension() function
  - isAllowedMimeType() function
- Refactor GetItemImageInteractor to use getMimeTypeFromFilename()
- Refactor UploadItemImagesInteractor to use isAllowedExtension() and isAllowedMimeType()
- Remove duplicate constants and validation logic

This refactoring centralizes image format validation logic,
making it easier to maintain and ensuring consistency across
the codebase.
